Refactor types file
- Split up to seperate files in the generator folder
- Remove unused types
