+++ R.swift/Generator/Typealias.swift
@@ -1,0 +1,20 @@
+//
+//  Typealias.swift
+//  R.swift
+//
+//  Created by Mathijs Kadijk on 10-12-15.
+//  Copyright © 2015 Mathijs Kadijk. All rights reserved.
+//
+
+import Foundation
+
+struct Typealias: CustomStringConvertible {
+  let alias: Type
+  let type: Type?
+
+  var description: String {
+    let typeString = type.map { " = \($0)" } ?? ""
+
+    return "typealias \(alias)\(typeString)"
+  }
+}
