+++ R.swift/func.swift
@@ -104,7 +104,7 @@ func imageStructFromAssetFolders(assetFolders: [AssetFolder], andImages images: 
     warn("Skipping \(duplicate.count) images because symbol '\(duplicate.first!.callName)' would be generated for all of these images: \(names)")
   }
 
-  return Struct(type: Type(name: "image"), lets: [], vars: vars.uniques, functions: [], structs: [])
+  return Struct(type: Type(name: "image"), vars: vars.uniques, functions: [], structs: [])
 }
 
 // Segue
@@ -164,14 +164,13 @@ func segueStructFromStoryboards(storyboards: [Storyboard]) -> Struct {
 
       return Struct(
         type: Type(name: sanitizedSwiftName(sourceType.description)),
-        lets: [],
         vars: vars,
         functions: [],
         structs: []
       )
     }
 
-  return Struct(type: Type(name: "segue"), lets: [], vars: [], functions: [], structs: structs)
+  return Struct(type: Type(name: "segue"), vars: [], functions: [], structs: structs)
 }
 
 // Storyboard
@@ -185,7 +184,7 @@ func storyboardStructAndFunctionFromStoryboards(storyboards: [Storyboard]) -> (S
   }
 
   return (
-    Struct(type: Type(name: "storyboard"), lets: [], vars: [], functions: [], structs: groupedStoryboards.uniques.map(storyboardStructForStoryboard)),
+    Struct(type: Type(name: "storyboard"), vars: [], functions: [], structs: groupedStoryboards.uniques.map(storyboardStructForStoryboard)),
     validateAllFunctionWithStoryboards(groupedStoryboards.uniques)
   )
 }
@@ -220,7 +219,7 @@ func storyboardStructForStoryboard(storyboard: Storyboard) -> Struct {
     }
   let validateViewControllersFunc = Function(isStatic: true, name: "validateViewControllers", generics: nil, parameters: [], returnType: Type._Void, body: validateViewControllersLines.joinWithSeparator("\n"))
 
-  return Struct(type: Type(name: sanitizedSwiftName(storyboard.name)), lets: [], vars: instanceVars + initialViewControllerVar + viewControllerVars, functions: [validateImagesFunc, validateViewControllersFunc], structs: [])
+  return Struct(type: Type(name: sanitizedSwiftName(storyboard.name)), vars: instanceVars + initialViewControllerVar + viewControllerVars, functions: [validateImagesFunc, validateViewControllersFunc], structs: [])
 }
 
 func validateAllFunctionWithStoryboards(storyboards: [Storyboard]) -> Function {
@@ -244,8 +243,8 @@ func nibStructFromNibs(nibs: [Nib]) -> (intern: Struct, extern: Struct) {
   }
 
   return (
-    intern: Struct(type: Type(name: "nib"), lets: [], vars: [], functions: [], structs: groupedNibs.uniques.map(nibStructForNib)),
-    extern: Struct(type: Type(name: "nib"), lets: [], vars: groupedNibs.uniques.map(nibVarForNib), functions: [], structs: [])
+    intern: Struct(type: Type(name: "nib"), vars: [], functions: [], structs: groupedNibs.uniques.map(nibStructForNib)),
+    extern: Struct(type: Type(name: "nib"), vars: groupedNibs.uniques.map(nibVarForNib), functions: [], structs: [])
   )
 }
 
@@ -290,7 +289,7 @@ func nibStructForNib(nib: Nib) -> Struct {
     parameters: instantiateParameters,
     returnType: Type(name: "[AnyObject]"),
     body: "return instance.instantiateWithOwner(ownerOrNil, options: optionsOrNil)"
-  ) as Func
+  )
 
   let viewFuncs = zip(nib.rootViews, Ordinals)
     .map { (view: $0.0, ordinal: $0.1) }
@@ -302,7 +301,7 @@ func nibStructForNib(nib: Nib) -> Struct {
         parameters: instantiateParameters,
         returnType: $0.view.asOptional(),
         body: "return \(instantiateFunc.callName)(ownerOrNil, options: optionsOrNil)[\($0.ordinal.number - 1)] as? \($0.view)"
-      ) as Func
+      )
     }
 
   let reuseIdentifierVars: [Var]
@@ -328,7 +327,6 @@ func nibStructForNib(nib: Nib) -> Struct {
     type: Type(name: "_\(sanitizedName)"),
     implements: [Type.NibResourceProtocol] + reuseProtocols,
     typealiasses: reuseTypealiasses,
-    lets: [],
     vars: [bundleVar, nameVar, instanceVar] + reuseIdentifierVars,
     functions: [instantiateFunc] + viewFuncs,
     structs: []
@@ -349,7 +347,7 @@ func reuseIdentifierStructFromReusables(reusables: [Reusable]) -> Struct {
     .uniques
     .map(varFromReusable)
 
-  return Struct(type: Type(name: "reuseIdentifier"), lets: [], vars: reuseIdentifierVars, functions: [], structs: [])
+  return Struct(type: Type(name: "reuseIdentifier"), vars: reuseIdentifierVars, functions: [], structs: [])
 }
 
 func varFromReusable(reusable: Reusable) -> Var {
@@ -364,7 +362,7 @@ func varFromReusable(reusable: Reusable) -> Var {
 // Fonts
 
 func fontStructFromFonts(fonts: [Font]) -> Struct {
-  return Struct(type: Type(name: "font"), lets: [], vars: [], functions: fonts.map(fontFunctionFromFont), structs: [])
+  return Struct(type: Type(name: "font"), vars: [], functions: fonts.map(fontFunctionFromFont), structs: [])
 }
 
 func fontFunctionFromFont(font: Font) -> Function {
@@ -394,7 +392,7 @@ func resourceStructFromResourceFiles(resourceFiles: [ResourceFile]) -> Struct {
     .uniques
     .map(varFromResourceFile)
 
-  return Struct(type: Type(name: "file"), lets: [], vars: resourceVars, functions: [], structs: [])
+  return Struct(type: Type(name: "file"), vars: resourceVars, functions: [], structs: [])
 }
 
 func varFromResourceFile(resourceFile: ResourceFile) -> Var {
